DAVIS: start moving common functionality between USB and RPi out to common header.
